fixup! Address PR comments.
